added mobile drawer for settings
Settings drawer is now swipeable on mobile devices.
