+++ src/components/profile/SettingsItem.jsx
@@ -10,10 +10,7 @@ import ListItemText from '@material-ui/core/ListItemText';
 const styles = theme => ({
   menuItem: {
     '&:focus': {
-      backgroundColor: theme.palette.primary.main,
-      '& $primary, & $icon': {
-        color: theme.palette.common.white,
-      },
+      backgroundColor: '#82B1FF',
     },
   },
 });
